refactor: move useAuth hook to its file
